moved IPopupMessage inside AbstractPopupMessage.hpp
