+++ Client/Client.Qt/Widgets/AbstractPopupMessage.cpp
@@ -1,0 +1,28 @@
+#include "AbstractPopupMessage.hpp"
+
+AbstractPopupMessage::AbstractPopupMessage(QWidget* parent) 
+	: QWidget(parent) 
+{
+}
+
+void AbstractPopupMessage::setPopupText(const QString& text) 
+{
+
+}
+
+void AbstractPopupMessage::popupShow() 
+{
+
+}
+
+float AbstractPopupMessage::getPopupOpacity() const 
+{ 
+	return _popupOpacity; 
+}
+
+void AbstractPopupMessage::setPopupOpacity(float opacity)
+{
+    _popupOpacity = opacity;
+
+    setWindowOpacity(opacity);
+}
